fix: move rendering to separate conditional block
Input handling and game play loop were tied together. To fix this I
separated the rendering of the game loop to only occur if playLoop is
true. This kept the input handling tied to the framerate, allowing for
keyboard events to be captured and stored, buffering inputs, before the
next frame is drawn.