fix: fix ball collision with bar
The collision check for the ball and bar was not working correctly. The
bounds defined in the collision check function overlapped 1 real pixel
right and down, making it so if the ball was one game pixel right or
down, collision could still occur.

Fixed by not including that extra pixel in the collision check function.